Ultimate Trivia Alpha Version
Replaced HighScore and Progress saving activity by using a new variable called 'uniqueQuizIdentifier'. This is a string that contains essentially a path to the exact quiz we are in. This is used for storing and retrieving progress and highscores. Code still needs to be cleaned up to get rid of old method variables such as 'subTopicName'